refactor: Optimizing the Card component
