+++ src/containers/Dashboard/NetworkStatistics/NetworkStatistics.tsx
@@ -2,7 +2,7 @@ import { Box, Typography } from '@mui/material'
 import numeral from 'numeral'
 import { useSelector } from 'react-redux'
 import { RootState } from 'store'
-import Card from 'components/Card/Card'
+import Card from 'components/Card'
 import { useOnlineVotingPower } from '../VotingPower/hooks'
 import { useDataBlocks } from './Blocks/hooks'
 import { useMarketRecoil } from './Market/hooks'
