+++ src/containers/Dashboard/WalletInformation/WalletInformation.tsx
@@ -4,7 +4,7 @@ import { useSelector } from 'react-redux'
 import { useNavigate } from 'react-router-dom'
 import { RootState } from 'store'
 import { copyToClipboard, formatAddress } from 'utils/projectUtils'
-import Card from 'components/Card/Card'
+import Card from 'components/Card'
 import TestAvatar from 'assets/vectors/test-avatar.svg'
 import CopyIcon from 'assets/vectors/copy-icon.svg'
 import EditIcon from 'assets/vectors/edit-icon.svg'
